Created a block meter
A block meter has been added to prevent the player from exploiting the block functionality. At this current point, the block meter only depletes as the player continues blocking and still allows the player to block even after the meter has fully depleted. More functionality to this bar will be added in a later commit.