Updated Crimson Meter
Crimson meter now checks if the value is above 0, if not then the player will not trigger any animations and instead a message will be pushed saying the player has no crimson. Further code is currently in development to change the overall functionality of the crimson meter, and test scene has been re-ordered to test out the crimson functionality.